Experiment_Code/requirements.txt
# THIS FILE CONTAILS THE REQUIRED LIBRARIES TO RUN THE CODE PROVIDED IN THIS REPOSITORY

pandas
numpy
PyQt5
PySimpleGUI
pyserial
typing
openpyxl
